+++ src/components/Hashtag/styles.ts
