+++ src/components/Loader/styles.ts
